refactor: move stuff to log
